move endGame from Main to Statuspanel  < organisation
